ten commit

+++ actor/editActor.php
@@ -10,25 +10,39 @@ if (!$conection) {
 
 $id = intval($_POST['idActor']);
 $nombre = $_POST['nombre'];
-$obrasFinal = $_POST['obrasFinal'];
+$obrasAdd = $_POST['obrasAdd'] != "" ? explode(",", $_POST['obrasAdd']) : "";
+$obrasDelete = $_POST['obrasDelete'] != "" ? explode(",", $_POST['obrasDelete']) : "";
 $sqlUpdate = "UPDATE actor SET nombre = '$nombre' WHERE id = $id";
-$arrayObras = explode(",", $obrasFinal);
 
 // consultas DELETE
 $sqlDelete = "";
-foreach ($arrayObras as $obra) {
-    $sqlDelete .= "DELETE FROM video_actor WHERE actor = $id AND video = $obra; ";
+if ($obrasDelete != "") {
+    foreach ($obrasDelete as $obra) {
+        $sqlDelete .= "DELETE FROM video_actor WHERE actor = $id AND video = $obra; ";
+    }
 }
 
 // consulta INSERT
 $sqlInsert = "INSERT INTO video_actor (actor, video) VALUES ";
 $values = [];
-foreach ($arrayObras as $obra) {
-    $values[] = "($id, $obra)";
+if ($obrasAdd != "") {
+    foreach ($obrasAdd as $obra) {
+        $values[] = "($id, $obra)";
+    }
+    $sqlInsert .= implode(", ", $values);
 }
-$sqlInsert .= implode(", ", $values);
 
-if ($conection->query($sqlUpdate) && $conection->query($sqlDelete) && $conection->query($sqlInsert)) {
+if ($conection->query($sqlUpdate)) {
+    // Si hay una consulta DELETE, ejecutarla
+    if ($sqlDelete != "") {
+        $conection->query($sqlDelete);
+    }
+
+    // Si hay una consulta INSERT, ejecutarla
+    if ($sqlInsert != "") {
+        $conection->query($sqlInsert);
+    }
+
     $_SESSION['alert'] = 'Se editó correctamente el actor';
     header('Location: ../home/home.view.php');
     exit();
